requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
pytz>=2025.1

